Attempt to clean up some of the files.
