+++ js/helpers/index.js
@@ -1,0 +1,5 @@
+import * as data from './data-helpers';
+import * as local from './local-helpers';
+
+export {data as data}
+export {local as local}
