+++ js/initialStates/data-state.js
@@ -1,0 +1,10 @@
+import Immutable from 'immutable';
+
+export default Immutable.fromJS({
+  data: [],
+  pageProperties: {
+    currentPage: 0,
+    maxPage: 0,
+    pageSize: 0
+  }
+});
