+++ js/initialStates/index.js
@@ -1,0 +1,5 @@
+import {default as data} from './data-state';
+import {default as local} from './local-state';
+
+export {data as data};
+export { local as local };
