+++ js/initialStates/local-state.js
@@ -1,0 +1,12 @@
+import Immutable from 'immutable';
+
+export default Immutable.fromJS({
+  pageProperties: {
+    pageSize: 10,
+    currentPage: 1,
+    sortColumns: [],
+    sortAscending: true
+  },
+  filter: '',
+  filteredData: Immutable.fromJS([])
+});
